streamlit/requirements.txt
streamlit==1.28.1
snowflake-connector-python==3.5.0
pandas==2.1.3
plotly==5.17.0
python-dotenv==1.0.0
